1.20.1 peatral mail fix (#203)
* Fix broken trader and catalogue

* Fix condense stacks

This fixes condenseStacks in a quite terrible way.
This changes the behaviour of containsSets that it
actually will count all sets, not only the largest
amount possible with one stack.

* Improve postage algorithm

Before it was possible that no combination of stamps was found due to the perfect combination not being available and no larger stamp being available.
If that's the case it will now sum up the largest stamps in order to try to match the required postage.
This is basically a greedy implementation. If this should be further optimized you should look into the knapsack problem.

* Update pobox info gui to toast

* Fix broken translation key

* Close #116

* Use config for pobox toast

* Move mail carriers to custom registry

* Fix mailbox toast typo

* Fix mail toast shown through player action

* Fix processed letter being editable

* Extract TradeRegistry into SavedData

* Save TradeStations via TradeRegistry

* Extract POBoxRegistry into SavedData

* Save POBoxes via POBoxRegistry

* Remove unnecessary caching of PostOffice

* Cleanup mail API

* Cleanup mail package

* Migrate EventHandlerMailAlert to ModuleMail

* Port changes to 1.20.1 (untested)

* Fix janky trade station naming GUI

* Improved textboxes + player address autocomplete for letter GUI

* Make catalogue and letters more intuitive

* Revert "Fix broken translation key"

This reverts commit 56718f5503c22dcf4223726acd3f1eb0f78af8c1.

* Fix language changes from bad cherry-pick

---------

Co-authored-by: Christof Reimers <contact@peatral.xyz>